Improve handling of image load delay on first page of search results.
This uses the index of the result to determine if it's on the first page, instead of the search offset.
