+++ src/components/search/result/SearchResultPanelContainer.js
@@ -9,7 +9,6 @@ import {
 import {
   getSearchError,
   getSearchNextOffset,
-  getSearchOffset,
   getSearchResult,
   isSearchPending,
 } from '../../../reducers';
@@ -18,7 +17,6 @@ const mapStateToProps = (state) => ({
   error: getSearchError(state),
   isPending: isSearchPending(state),
   nextOffset: getSearchNextOffset(state),
-  offset: getSearchOffset(state),
   result: getSearchResult(state),
 });
